MGSim: lower autoconf requirement to 2.61 which is reasonable and pleases Ubuntu users. Add include <cstring> to main.cpp for strrchr. Use 'double' for FPU::QueueOperation to gain a few more bits of precision.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@1858 e97e5017-a994-416e-809d-76780e9f78db

+++ src/FPU.cpp
@@ -8,7 +8,7 @@ using namespace std;
 namespace Simulator
 {
 
-bool FPU::QueueOperation(FPUOperation op, int size, float Rav, float Rbv, const RegAddr& Rc)
+bool FPU::QueueOperation(FPUOperation op, int size, double Rav, double Rbv, const RegAddr& Rc)
 {
 	CycleNo	latency = 0;
 	double  value;
